refactor: rework admin's sudo login
- simplify the implementation
- rename the middleware to "sudo"
- renew the session on each request
- lower the session timeout to 5 minutes
